delete warning added

+++ app/static/js/controllers/tournament-controller.js
@@ -3,13 +3,15 @@ App.TournamentController = Ember.ObjectController.extend({
         removeTournament: function() {
             var tournament = this.get('model');
 
-            var onFail = function(response) {
-                alert(response.responseText);
-                tournament.rollback();
-            };
+            if (confirm("Are you sure? There is no undelete?")) {
+                var onFail = function(response) {
+                    alert(response.responseText);
+                    tournament.rollback();
+                };
 
-            tournament.deleteRecord();
-            tournament.save().then(function() {}, onFail);
+                tournament.deleteRecord();
+                tournament.save().then(function() {}, onFail);
+            }
         }
     }
 });
